requirements.txt
langchain
langchain_community
langchain_qdrant
sentence_transformers==2.7.0
langchain_huggingface
streamlit
qdrant-client
langchain_groq
PyPDF2
python-dateutil